upgraded to gitlab4j5.1.0

+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/GitLabProjectApi.java
@@ -273,7 +273,7 @@ public class GitLabProjectApi extends GitLabApiWithFileAccess implements Project
 
     private boolean tryEnsureMinAccessLevel(GitLabApi gitLabApi, org.gitlab4j.api.models.Project project, AccessLevel accessLevel)
     {
-        int projectId = project.getId();
+        long projectId = project.getId();
 
         User currentUser;
         try
@@ -288,7 +288,7 @@ public class GitLabProjectApi extends GitLabApiWithFileAccess implements Project
         }
 
         org.gitlab4j.api.ProjectApi projectApi = gitLabApi.getProjectApi();
-        int userId = currentUser.getId();
+        long userId = currentUser.getId();
         CallUntil<AccessLevel, ?> callUntil = CallUntil.callUntil(() ->
         {
             try
@@ -340,7 +340,7 @@ public class GitLabProjectApi extends GitLabApiWithFileAccess implements Project
 
         // Get project ID
         GitLabProjectId projectId = id.chars().allMatch(Character::isDigit) ?
-                GitLabProjectId.newProjectId(getGitLabConfiguration().getProjectIdPrefix(), Integer.parseInt(id)) :
+                GitLabProjectId.newProjectId(getGitLabConfiguration().getProjectIdPrefix(), Long.parseLong(id)) :
                 parseProjectId(id);
 
         // Find project
