+++ legend-sdlc-server/src/test/java/org/finos/legend/sdlc/server/gitlab/api/GitLabEntityApiTestResource.java
@@ -161,8 +161,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        int parsedMergeRequestId = Integer.parseInt(reviewId);
-        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        long parsedMergeRequestId = Long.parseLong(reviewId);
+        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
@@ -308,8 +308,8 @@ public class GitLabEntityApiTestResource
 
         GitLabProjectId sdlcGitLabProjectId = GitLabProjectId.parseProjectId(projectId);
         MergeRequestApi mergeRequestApi = gitLabMemberUserContext.getGitLabAPI().getMergeRequestApi();
-        int parsedMergeRequestId = Integer.parseInt(reviewId);
-        int gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
+        long parsedMergeRequestId = Long.parseLong(reviewId);
+        long gitlabProjectId = sdlcGitLabProjectId.getGitLabId();
 
         String requiredStatus = "can_be_merged";
         CallUntil<MergeRequest, GitLabApiException> callUntil = CallUntil.callUntil(
